blogapi/requirements.txt
asgiref==3.5.2
certifi @ file:///C:/b/abs_ac29jvt43w/croot/certifi_1665076682579/work/certifi
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
defusedxml==0.7.1
Django==4.1.3
django-allauth==0.51.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.14.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.2
openapi-codec==1.3.2
psycopg2==2.9.5
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.12
wincertstore==0.2
